Redefined page_index_t to intptr_t, for mswin/amd64's sake

+++ src/runtime/gc.h
@@ -21,7 +21,7 @@
 
 #define PAGE_BYTES BACKEND_PAGE_BYTES
 
-typedef signed long page_index_t;
+typedef intptr_t page_index_t;
 #define PAGE_INDEX_FMT "ld"
 
 typedef signed char generation_index_t;
